+++ x-pack/plugin/core/src/main/java/org/elasticsearch/xpack/core/dataframe/DataFrameFeatureSetUsage.java
@@ -4,7 +4,7 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-package org.elasticsearch.xpack.ml.featureindexbuilder;
+package org.elasticsearch.xpack.core.dataframe;
 
 import org.elasticsearch.common.io.stream.StreamInput;
 import org.elasticsearch.xpack.core.XPackField;
@@ -12,12 +12,12 @@ import org.elasticsearch.xpack.core.XPackFeatureSet.Usage;
 
 import java.io.IOException;
 
-public class FeatureIndexBuilderFeatureSetUsage extends Usage {
-    public FeatureIndexBuilderFeatureSetUsage(StreamInput input) throws IOException {
+public class DataFrameFeatureSetUsage extends Usage {
+    public DataFrameFeatureSetUsage(StreamInput input) throws IOException {
         super(input);
     }
 
-    public FeatureIndexBuilderFeatureSetUsage(boolean available, boolean enabled) {
+    public DataFrameFeatureSetUsage(boolean available, boolean enabled) {
         super(XPackField.DATA_FRAME, available, enabled);
     }
 }
